Moved Models into own namespace
