+++ src/Response/Model/Wallet/History.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Electrum\Response\Wallet;
+namespace Electrum\Response\Model\Wallet;
 
 use Electrum\Response\ResponseInterface;
 use Electrum\Response\Traits\Transactions;
